+++ changelog.md
@@ -1,17 +1,24 @@
 # Changelog
 
+## 0.2.2 - 2023-01-10
+
+### Added
+
+-   Button to deactivate offline access
+-   Messages indicating whether an action was successful or not
+
 ## 0.1.2 - 2023-01-10
 
 ### Changed
 
-- Added vendor prefix to module name
-- Improved module meta
+-   Added vendor prefix to module name
+-   Improved module meta
 
 ## 0.1.1 - 2022-12-20
 
 ### Fixed
 
-- Minor fixes
+-   Minor fixes
 
 ## 0.1.0 - 2022-12-20